server/requirements.txt
django
python-dotenv
openai
numpy
faiss-cpu
pypdf2
langchain
langchain-openai